Put the finishing touches to /admin/users
