+++ src/components/data-display/User.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
-  import { Avatar, popup } from '@skeletonlabs/skeleton';
+  import { Avatar } from '@skeletonlabs/skeleton';
   import { buildUrl } from 'osu-web.js';
-  import type { PopupSettings } from '@skeletonlabs/skeleton';
 
   export let data: {
     user: {
@@ -18,20 +17,10 @@
         code: string
       }
     };
-    options: {
-      forceShowDiscord: boolean,
-      management?: {
-        adminChange: Function,
-        deleteUser: Function
-      }
+    options?: {
+      forceShowDiscord?: boolean
     };
   };
-
-  const popupUserManage: PopupSettings = {
-    event: 'click',
-    target: 'popupUserManage',
-    placement: 'bottom'
-  }
 </script>
 
 <div class="card w-64 p-4 bg-surface-backdrop-token">
@@ -56,7 +45,7 @@
     <div class="col-span-1">
     </div>
     <div class="col-span-3">
-      {#if data.user.showDiscordTag || data.options.forceShowDiscord}
+      {#if data.user.showDiscordTag || data.options?.forceShowDiscord}
         {data.user.discordUsername}#{data.user.discordDiscriminator}
       {/if}
     </div>
@@ -68,17 +57,5 @@
         <span class="badge variant-filled-error mx-auto">Restricted</span>
       {/if}
     </div>
-    {#if data.options.management !== undefined}
-      <div class="mt-2 col-span-4 row-span-4">
-        <button class="mx-auto btn variant-filled-tertiary" use:popup={popupUserManage}>Manage User...</button>
-        <div class="card p-4 max-w-sm" data-popup="popupUserManage">
-          <div class="grid grid-cols-1 gap-2">
-            <button class="btn variant-filled-error" on:click={() => {data.options.management?.adminChange(data.user)}}>{data.user.isAdmin ? "Remove" : "Add"} Admin</button>
-            <button class="btn variant-filled-error" on:click={() => {data.options.management?.deleteUser(data.user)}}>Delete User</button>
-          </div>
-          <div class="arrow bg-surface-100-800-token" />
-        </div>
-      </div>
-    {/if}
   </div>
 </div>
